requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
geopy==2.4.1
python-dateutil==2.8.2